pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lubrication-solver"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "ALLOUCHE", email = "david_allouche@outlook.com"}
]
dependencies = [
    "numpy>=2.0.0",
    "pyyaml>=6.0.3",
    "matplotlib>=3.5.0"
]

[project.optional-dependencies]
dev = [
    "black>=22.0",
    "pytest>=7.0",
]

[project.scripts]
lubrication-solver = "lubrication_solver.cli:main"

[tool.black]
line-length = 88
target-version = ['py36']
skip-string-normalization = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
